+++ liquepy/__init__.py
@@ -14,3 +14,4 @@ try:
     from liquepy import spatial
 except ImportError:
     pass
+from . import esp